added authorization for professor and fixed profesors tests

+++ controllers/professors/professorController.js
@@ -54,6 +54,8 @@ const deleteProfessor = async (req, res) => {
     
   try {
     let ID = req.params.id;
+    if(req.userData.id != ID)
+      return res.sendStatus(STATUS_CODES.FORBIDDEN);
     const professor = await Professor.destroy({
       where: {id: ID},
     });
@@ -71,6 +73,8 @@ const deleteProfessor = async (req, res) => {
 const updateProfessor = async (req, res) => {
   try {
     let ID = req.params.id;
+    if(req.userData.id != ID)
+      return res.sendStatus(STATUS_CODES.FORBIDDEN);
     const hash = bcrypt.hashSync(req.body.password, +process.env.BCRYPT_SALT_ROUNDS);
     const professor = await Professor.update({
       name: req.body.name,
